Workaround and checkstyle complaint
