+++ src/main/java/io/github/fablabsmc/fablabs/api/gamerule/v1/rule/FloatRule.java
@@ -1,9 +1,10 @@
-package com.martmists.libgamerule.api.rule;
+package io.github.fablabsmc.fablabs.api.gamerule.v1.rule;
 
-import com.martmists.libgamerule.api.FloatSupplier;
-import com.martmists.libgamerule.api.RuleFactory;
-import com.martmists.libgamerule.impl.GameRuleRegistryImpl;
 import com.mojang.brigadier.context.CommandContext;
+import io.github.fablabsmc.fablabs.api.gamerule.v1.FloatSupplier;
+import io.github.fablabsmc.fablabs.api.gamerule.v1.RuleFactory;
+import io.github.fablabsmc.fablabs.impl.gamerule.GameRuleRegistryImpl;
+
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.world.GameRules;
 
@@ -13,7 +14,6 @@ public class FloatRule extends GameRules.Rule<FloatRule> implements FloatSupplie
 	// TODO: i509VCB - Should we make these constructors private since people are not supposed to be able to invoke these, and then use some invokers to create these internally within the api?
 
 	/**
-	 * @param type the rule type
 	 * @deprecated Please use {@link RuleFactory} instead.
 	 */
 	@Deprecated
