+++ src/main/java/io/github/fablabsmc/fablabs/api/gamerule/v1/rule/TextRule.java
@@ -1,40 +1,41 @@
-package com.martmists.libgamerule.api.rule;
-
-import com.martmists.libgamerule.api.RuleFactory;
-import com.mojang.brigadier.context.CommandContext;
-import net.minecraft.server.command.ServerCommandSource;
-import net.minecraft.world.GameRules;
+package io.github.fablabsmc.fablabs.api.gamerule.v1.rule;
 
 import java.util.function.Supplier;
 
-public class StringRule extends GameRules.Rule<StringRule> implements Supplier<String> {
-	private String value;
+import com.mojang.brigadier.context.CommandContext;
+import io.github.fablabsmc.fablabs.api.gamerule.v1.RuleFactory;
+
+import net.minecraft.server.command.ServerCommandSource;
+import net.minecraft.text.Text;
+import net.minecraft.world.GameRules;
+
+public class TextRule extends GameRules.Rule<TextRule> implements Supplier<Text> {
+	private Text value;
 
 	// TODO: i509VCB - Should we make these constructors private since people are not supposed to be able to invoke these, and then use some invokers to create these internally within the api?
 
 	/**
-	 * @param type the rule type
 	 * @deprecated Please use {@link RuleFactory} instead.
 	 */
 	@Deprecated
-	public StringRule(GameRules.RuleType<StringRule> type, String value) {
+	public TextRule(GameRules.RuleType<TextRule> type, Text defaultValue) {
 		super(type);
-		this.value = value;
+		this.value = defaultValue;
 	}
 
 	@Override
 	protected void setFromArgument(CommandContext<ServerCommandSource> context, String name) {
-		this.value = context.getArgument(name, String.class);
+		this.value = context.getArgument(name, Text.class);
 	}
 
 	@Override
 	protected void deserialize(String value) {
-		this.value = value;
+		this.value = Text.Serializer.fromJson(value);
 	}
 
 	@Override
 	protected String serialize() {
-		return this.value;
+		return Text.Serializer.toJson(this.value);
 	}
 
 	@Override
@@ -43,12 +44,12 @@ public class StringRule extends GameRules.Rule<StringRule> implements Supplier<S
 	}
 
 	@Override
-	protected StringRule getThis() {
+	protected TextRule getThis() {
 		return this;
 	}
 
 	@Override
-	public String get() {
+	public Text get() {
 		return this.value;
 	}
 }
