+++ src/main/java/io/github/fablabsmc/fablabs/impl/gamerule/EnumArgumentType.java
@@ -1,4 +1,9 @@
-package com.martmists.libgamerule.arguments;
+package io.github.fablabsmc.fablabs.impl.gamerule;
+
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.concurrent.CompletableFuture;
+import java.util.stream.Collectors;
 
 import com.mojang.brigadier.StringReader;
 import com.mojang.brigadier.arguments.ArgumentType;
@@ -7,13 +12,9 @@ import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import com.mojang.brigadier.exceptions.SimpleCommandExceptionType;
 import com.mojang.brigadier.suggestion.Suggestions;
 import com.mojang.brigadier.suggestion.SuggestionsBuilder;
+
 import net.minecraft.server.command.CommandSource;
 import net.minecraft.text.LiteralText;
-
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.concurrent.CompletableFuture;
-import java.util.stream.Collectors;
 
 public class EnumArgumentType<E extends Enum<E>> implements ArgumentType<E> {
 	Class<E> clazz;
