+++ src/main/java/io/github/fablabsmc/fablabs/mixin/gamerule/GameRuleCommandMixin$RuleConsumer.java
@@ -1,15 +1,16 @@
-package com.martmists.libgamerule.mixin;
+package io.github.fablabsmc.fablabs.mixin.gamerule;
 
-import com.martmists.libgamerule.impl.LiteralRuleCommand;
-import com.martmists.libgamerule.impl.LiteralRuleType;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
-import net.minecraft.server.command.ServerCommandSource;
-import net.minecraft.world.GameRules;
+import io.github.fablabsmc.fablabs.impl.gamerule.LiteralRuleCommand;
+import io.github.fablabsmc.fablabs.impl.gamerule.LiteralRuleType;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import net.minecraft.server.command.ServerCommandSource;
+import net.minecraft.world.GameRules;
 
 @Mixin(targets = "net/minecraft/server/command/GameRuleCommand$1")
 public abstract class GameRuleCommandMixin$RuleConsumer {
