+++ src/main/java/io/github/fablabsmc/fablabs/mixin/gamerule/GameRulesAccessor.java
@@ -1,8 +1,9 @@
-package com.martmists.libgamerule.mixin;
+package io.github.fablabsmc.fablabs.mixin.gamerule;
 
-import net.minecraft.world.GameRules;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Invoker;
+
+import net.minecraft.world.GameRules;
 
 @Mixin(GameRules.class)
 public interface GameRulesAccessor {
